rename : 목적에 맞는 위치 변경
