+++ graduate/src/pages/Notice/NoticeRow.jsx
